Rename rest/utils.js to rest/authFetch; refactor
