+++ test/unit/utils.test.js
@@ -1,6 +1,6 @@
 import assert from 'assert'
 import sinon from 'sinon'
-import { authFetch } from '../../src/rest/utils'
+import authFetch from '../../src/rest/authFetch'
 
 const express = require('express')
